third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGLineElement-dom-x2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x2' attribute of the SVGLineElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lineElement.getAttribute('x2') is "100"
PASS lineElement.getAttribute('x2') is "200"
PASS successfullyParsed is true

TEST COMPLETE
